# Perf: Directly use ImmutableDictionary for TaskItem metadata