--- a/src/Build.UnitTests/Instance/TaskItem_Tests.cs
+++ b/src/Build.UnitTests/Instance/TaskItem_Tests.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.Linq;
 using Microsoft.Build.BackEnd;
 using Microsoft.Build.Collections;
@@ -191,11 +192,12 @@ namespace Microsoft.Build.UnitTests.OM.Instance
             item.MetadataNames.Cast<string>().ShouldBeSetEquivalentTo(new[] { "a", "b" }.Concat(s_builtInMetadataNames));
             item.MetadataCount.ShouldBe(s_builtInMetadataNames.Length + 2);
             item.DirectMetadataCount.ShouldBe(1);
+            item.HasCustomMetadata.ShouldBeTrue();
 
-            ICopyOnWritePropertyDictionary<ProjectMetadataInstance> metadata = item.MetadataCollection;
+            ImmutableDictionary<string, string> metadata = item.MetadataCollection;
             metadata.Count.ShouldBe(2);
-            metadata["a"].EvaluatedValue.ShouldBe("override");
-            metadata["b"].EvaluatedValue.ShouldBe("base");
+            metadata["a"].ShouldBe("override");
+            metadata["b"].ShouldBe("base");
 
             item.EnumerateMetadata().ShouldBeSetEquivalentTo(new KeyValuePair<string, string>[] { new("a", "override"), new("b", "base") });
 
@@ -219,12 +221,12 @@ namespace Microsoft.Build.UnitTests.OM.Instance
                     }
                 }
 
-                CopyOnWritePropertyDictionary<ProjectMetadataInstance> directMetadata = new();
+                ImmutableDictionary<string, string> directMetadata = ImmutableDictionaryExtensions.EmptyMetadata;
                 if (metadata is not null)
                 {
                     foreach ((string name, string value) in metadata)
                     {
-                        directMetadata.Set(new(name, value));
+                        directMetadata = directMetadata.SetItem(name, value);
                     }
                 }