--- a/src/Build/BackEnd/Components/RequestBuilder/Lookup.cs
+++ b/src/Build/BackEnd/Components/RequestBuilder/Lookup.cs
@@ -8,6 +8,7 @@ using System.Linq;
 using Microsoft.Build.Collections;
 using Microsoft.Build.Evaluation;
 using Microsoft.Build.Execution;
+using Microsoft.Build.Framework;
 using Microsoft.Build.Shared;
 using ReservedPropertyNames = Microsoft.Build.Internal.ReservedPropertyNames;
 
@@ -855,39 +856,22 @@ namespace Microsoft.Build.BackEnd
         /// </summary>
         private static void ApplyMetadataModificationsToItem(MetadataModifications modificationsToApply, ProjectItemInstance itemToModify)
         {
+            // PERF: Avoid additional allocations by going through the interfaces - ProjectItemInstance hides some symbols
+            // from its public API.
+            ITaskItem taskItem = itemToModify;
+            IMetadataContainer metadataContainer = itemToModify;
+
             // Remove any metadata from the item which is slated for removal.  The indexer in the modifications table will
             // return a modification with Remove == true either if there is an explicit entry for that name in the modifications
             // or if keepOnlySpecified == true and there is no entry for that name.
-            if (modificationsToApply.KeepOnlySpecified)
+            if (modificationsToApply.KeepOnlySpecified && metadataContainer.HasCustomMetadata)
             {
-
-                // Perf: Avoid boxing when possible by getting the underlying struct enumerator if available.
-                List<string> metadataToRemove = new List<string>(); ;
-                if (itemToModify.Metadata is CopyOnWritePropertyDictionary<ProjectMetadataInstance> copyOnWritePropertyMetadata)
+                foreach (KeyValuePair<string, string> m in metadataContainer.BackingMetadata.Dictionary)
                 {
-                    foreach (var m in copyOnWritePropertyMetadata)
+                    if (modificationsToApply[m.Key].Remove)
                     {
-                        string name = m.Value.Name;
-                        if (modificationsToApply[name].Remove)
-                        {
-                            metadataToRemove.Add(name);
-                        }
+                        taskItem.RemoveMetadata(m.Key);
                     }
-                }
-                else
-                {
-                    foreach (var m in itemToModify.Metadata)
-                    {
-                        if (modificationsToApply[m.Name].Remove)
-                        {
-                            metadataToRemove.Add(m.Name);
-                        }
-                    }
-                }
-
-                foreach (var metadataName in metadataToRemove)
-                {
-                    itemToModify.RemoveMetadata(metadataName);
                 }
             }
 
@@ -896,11 +880,11 @@ namespace Microsoft.Build.BackEnd
             {
                 if (modificationPair.Value.Remove)
                 {
-                    itemToModify.RemoveMetadata(modificationPair.Key);
+                    taskItem.RemoveMetadata(modificationPair.Key);
                 }
                 else if (modificationPair.Value.NewValue != null)
                 {
-                    itemToModify.SetMetadata(modificationPair.Key, modificationPair.Value.NewValue);
+                    taskItem.SetMetadata(modificationPair.Key, modificationPair.Value.NewValue);
                 }
             }
         }