… src/Framework/MSBuildNameIgnoreCaseComparer.cs
… src/Framework/SerializableMetadata.cs
… src/MSBuildTaskHost/MSBuildTaskHost.csproj
… src/Shared/CopyOnWriteDictionary.cs
… src/Shared/TaskParameter.cs
… src/Shared/TranslatorHelpers.cs
… src/Shared/XMakeElements.cs
… src/Tasks/Microsoft.Build.Tasks.csproj
… src/Utilities/Microsoft.Build.Utilities.csproj
… src/Utilities/TaskItem.cs