### Fixes

Fairly complex refactor to replace `TaskItem` metadata wrapper classes with direct use of `ImmutableDictionary<string, string>`. This eliminates multiple categories of allocations and significantly reduces known hot paths such as `GatherTaskItemOutputs` and `CopyMetadataTo`.

On OrchardCore, this change alone cuts RAR by ~40% + many other tasks, reduces allocations by ~1.4GB, and has a noticeable impact on E2E build time.

Also fixes #11895

**GatherTaskItemOutputs - Before**
<img width="727" height="212" alt="image" src="https://github.com/user-attachments/assets/94b58e27-eca5-4950-90b9-4eacf89d3a84" />

**GatherTaskItemOutputs - After**
<img width="863" height="230" alt="image" src="https://github.com/user-attachments/assets/98335118-41f7-4c90-9033-b50e1c640f10" />

**/m:1 Total Allocations - Before**
<img width="287" height="185" alt="image" src="https://github.com/user-attachments/assets/b4c4f674-9f5a-4bb7-8979-68d124c9fde0" />

**/m:1 Total Allocations - After**
<img width="292" height="199" alt="image" src="https://github.com/user-attachments/assets/c692e8dd-be5a-4c8a-ab8f-9290c4185ce8" />

**/m:1 Working Set - Before**
<img width="330" height="230" alt="image" src="https://github.com/user-attachments/assets/409d79ec-bdd5-448c-9511-952d6bebba41" />

**/m:1 Working Set - After**
<img width="352" height="241" alt="image" src="https://github.com/user-attachments/assets/45e5714d-9edc-4415-a77b-a5dd325fe2dd" />

**/m cold incremental build (no nodes) - Before**
```
...
     2771 ms  ResolvePackageAssets                     436 calls
     3057 ms  ProcessFrameworkReferences               436 calls
     5804 ms  ResolvePackageFileConflicts              436 calls
     6904 ms  Copy                                     7037 calls
    25061 ms  ResolveAssemblyReference                 436 calls
    144029 ms  CallTarget                               872 calls
    7396229 ms  MSBuild                                  2388 calls

Build succeeded.
    0 Warning(s)
    0 Error(s)

Time Elapsed 00:00:42.91
```

**/m cold incremental build (no nodes) - After**
```
...
     1861 ms  ResolvePackageAssets                     436 calls
     2107 ms  AssignProjectConfiguration               436 calls
     2224 ms  ProcessFrameworkReferences               436 calls
     5492 ms  ResolvePackageFileConflicts              436 calls
     5770 ms  Copy                                     7037 calls
    16408 ms  ResolveAssemblyReference                 436 calls
    163157 ms  CallTarget                               872 calls
    5339115 ms  MSBuild                                  2388 calls

Build succeeded.
    0 Warning(s)
    0 Error(s)

Time Elapsed 00:00:34.41
```

**/m hot incremental build (existing nodes) - Before**
```
...
     1311 ms  ResolvePackageAssets                     436 calls
     1611 ms  ResolveTargetingPackAssets               436 calls
     5295 ms  ResolvePackageFileConflicts              436 calls
     5965 ms  Copy                                     7037 calls
    21149 ms  ResolveAssemblyReference                 436 calls
    76149 ms  CallTarget                               872 calls
    3354848 ms  MSBuild                                  2388 calls

Build succeeded.
    0 Warning(s)
    0 Error(s)

Time Elapsed 00:00:27.71
```

**/m hot incremental build (existing nodes) - After**

```
...
      782 ms  ProcessFrameworkReferences               436 calls
      897 ms  AssignProjectConfiguration               436 calls
      954 ms  Microsoft.Build.Tasks.Git.GetUntrackedFiles 420 calls
      955 ms  ResolvePackageAssets                     436 calls
     4440 ms  Copy                                     7037 calls
     4434 ms  ResolvePackageFileConflicts              436 calls
    12249 ms  ResolveAssemblyReference                 436 calls
    100407 ms  CallTarget                               872 calls
    3321973 ms  MSBuild                                  2388 calls

Build succeeded.
    0 Warning(s)
    0 Error(s)

Time Elapsed 00:00:23.32

```

### Context

`ProjectItemInstance.TaskItem` (engine side) and `Utilities.TaskItem` (task side) currently use a various wrapping collections to implement their backing dictionaries:

*ProjectItemInstance.TaskItem*
- `ProjectMetadataInstance` to wrap key/value pairs and validate keys
- `CopyOnWritePropertyDictionary<ProjectMetadataInstance>` to provide a COW API that automatically converts and enumerates over `ProjectMetadataInstance`

*Utilities.TaskItem*
- `CopyOnWriteDictionary<string, string>` to provide a COW API

There are a few main issues here:

- Given that these are the most allocated objects in MSBuild, even these basic wrapping instances are responsible for a large number of allocations.
- Because the types between `ImmutableDictionary` implementations don't match, "cloning" to another type always results in fully rebuilding the dictionary instance.
- `ProjectMetadataInstance` is just the key + evaluated value, and everything else is derived - so we're essentially doubling the required memory per-entry.

### Changes Made

The overall gist of this change is replacing the above objects with direct use of `ImmutableDictionary<string, string>`, and adding a number of fast-paths that take advantage of this.

- To replicate the reserved name + itemspec modifier validation in `ProjectMetadataInstance`, `ImmutableDictionaryExtensions.SetItems()` is used throughout the Engine side. Otherwise, these are really only used now in the public API which can't be changed.
- `ImmutableDictionaryExtensions.EmptyMetadata` provides a singleton with `MSBuildNameIgnoreCaseComparer` already configured.
- To handle cross-AppDomain calls via `IMetadataContainer`, `SerializableMetadata` replaces `CopyOnWriteDictionary` as a simple struct wrapper that handles serialization.
- Use of `ICopyOnWritePropertyDictionary` is removed from `ProjectItemInstance` and `ItemDefinitionInstance` since this would add a fair amount of complexity to extracting and type checking backing dictionaries.

### Testing Notes

One area that isn't well tested is the immutable instance caching that was added for the CPS evaluation cache (which I believe is used by C# Dev Kit and VS?). Since this change removes use of interfaces for backing dictionaries, this bypasses the converters and instance caches (see `ProjectInstance.cs`).

I **suspect** that the massive savings in allocations and CPU here would offset any gains made by the immutable instance caching for `ICopyOnWritePropertyDictionary` and `ProjectMetadataInstance`. Plus the higher-level classes such as `ProjectItemInstance` are still cached, so I'm not even sure if it makes sense to still apply caching to the backing fields.

Either way though, the main issue is that there doesn't seem to be any unit testing around that functionality.

There's also a decent amount of dead code now in the replaced classes and room for merging some APIs given we have a single concrete-type, but this PR is already big enough as-is 😅

### Other Motivation

I also have a prototype of an array-based immutable dictionary that cuts off even **more** time here, and additionally Translator serialization of TaskItems since the slowest area is serializing `ImmutableDictionary`. That would cut down a large chunk of overhead off of the TaskHost and RAR out-of-proc, and after this change, it would be essentially a drop-in replacement.

### Bonus allocation / perf data

**ImmutableDictionary + related - Before**
<img width="948" height="505" alt="image" src="https://github.com/user-attachments/assets/c3f638bf-b6f4-4965-b1e7-db5089d47185" />

**ImmutableDictionary + related - After**
<img width="875" height="526" alt="image" src="https://github.com/user-attachments/assets/46034d00-b913-4950-a1fd-e3990b0f1e5c" />

**CopyOnWriteDictionary - Before**
<img width="951" height="405" alt="image" src="https://github.com/user-attachments/assets/c6f26435-0cea-43fe-8cbb-dc3c2ecafc5b" />

**CopyOnWriteDictionary - After**
<img width="875" height="413" alt="image" src="https://github.com/user-attachments/assets/cfcc5df9-543a-4ea4-a892-606bcc0e39b1" />

**ProjectMetadataInstance - Before**
<img width="958" height="509" alt="image" src="https://github.com/user-attachments/assets/94fced0c-3302-4f7f-9909-a526a6484128" />

**ProjectMetadataInstance - After**
<img width="890" height="361" alt="image" src="https://github.com/user-attachments/assets/bf2bb957-73bf-4592-9d75-d8af7d5c2c92" />