@yuehuang010 2025-07-28
Nice, I wonder if this fixes my reported issue too.  [#10490](https://github.com/dotnet/msbuild/issues/10490)